없는 알림을 삭제 시 오류 없이 리턴하도록 변경

+++ server/src/notifications/commands/handlers/delete-notification.handler.ts
@@ -10,12 +10,13 @@ export class DeleteNotificationHandler
     private readonly notificationsRepository: NotificationsRepository,
   ) {}
 
-  async execute(command: DeleteNotificationCommand): Promise<any> {
+  async execute(command: DeleteNotificationCommand): Promise<boolean> {
     const { notificationId, receiverNickname } = command;
     const notification = await this.notificationsRepository.findOne({
       notificationId: { eq: notificationId },
       receiverNickname: { eq: receiverNickname },
     });
+    if (!notification) return false;
     return this.notificationsRepository.remove(notification);
   }
 }
